+++ NEWS.rst
@@ -14,6 +14,11 @@ Release date: *not yet released*
   to :py:meth:`Table.scan` (`issue #39
   <https://github.com/wbolster/happybase/issues/39>`_).
 
+* Allow `None` as a valid value for the `batch_size` argument to
+  :py:meth:`Table.scan`, since HBase does not support specifying a batch size
+  when some scanner filters are used. (`issue #54
+  <https://github.com/wbolster/happybase/issues/54>`_).
+
 
 HappyBase 0.7
 -------------